Make Other Article in Connection uneditable
This is to prevent possible weirdness that may result from being
able to change the connected Article from the EditConnectionActivity.
(e.g. setting the Connection to another Article who already has a
Connection to the main Article.)

The user can quickly delete the Connection and then make a new one,
so changing the connected Article isn't a feature one would need
anyway.