Use empty String to show no role in Membership
When a Membership is saved with an empty role via
EditMembershipActivity, the text file that is produced contains a
newline character.

When this text file is read, it returns an empty string instead of null
due to the newline character.

As such, in order for the role TextView to be properly hidden on
Membership cards, the card must check whether membership.memberRole is
empty, rather than null.