Grayscale sky blue-colored drawables
Since these drawables will be used across all themes, using a "neutral"
colour will help them fit in better.

Still, the gray is quite jarring compared to the two-tone colour scheme
of the rest of the app, so in the future, some way of colour-coding each
asset to the current theme may be included.