Add a "dummy" layout to space out BottomBar
The plan is to only display the currently-selected Category's name
on the BottomBar. This means that the other TextViews should have their
visibility set to GONE... but this comes with the side-effect of
shrinking each non-active button's height, making the background color
"strip" have uneven height.

To remedy this, a "dummy" layout has been placed that contains just one
button and TextView, both of which are set to INVISIBLE (unseen but
still taking up space). The background color has also been migrated to
the root layout, whose height is now fixed to match that of the "dummy"
layout at all times.