Fix bug where Articles cannot be deleted immediately after opening them
This bug occurred because the change in a781511455c8f089428bad18340657069b26c736 causes folders to be remade whenever the Article Activity's onPause function is called. The onPause calls functions to save Article files, but only if the respective 'wasChanged' variable is set to True.

When the Article Activity is opened, the text fields are edited to display previously-saved content. This triggers their onChange event, which we've configured to set the aforementioned 'wasChanged' variable to true.

The solution is to only set this 'wasChanged' variable to true if the field was edited by the user (and don't do anything if it was edited programmatically).