Convert file URIs to content URIs for Android versions >= 10
Based on the investigation in #43, the external storage issue seems to
occur when the file root URI is a "file:///" URI on an Android 10+
device.  As such, the solution may be to convert these URIs to
"content:///" URIs.

For cases that don't match the criteria above (the URI is already
a content URI, or Android version < 10), we don't do any conversion.
This is to maintain stability for users unaffected by this issue.