Display Announcements immediately on startup
Due to breaking changes in Android 13+, the app might no longer function
on some devices. By displaying the announcements as early as possible on
app launch, we allow those affected to view the final announcement and
thus see an explanation for why the app is no longer working, find out
how to recover their World data files, and discover the installation
page for World Scribe 2.